Combine report unknown and missing marker function
