+++ lib/markout/index.js
@@ -13,8 +13,7 @@ var constructVideos = require("./construct-videos");
 var constructTables = require("./construct-tables");
 var constructReveals = require("./construct-reveals");
 var constructComments = require("./construct-comments");
-var missingMarkers = require("./report-missing-markers");
-var unknownMarkers = require("./report-unknown-markers");
+var invalidMarkers = require("./report-invalid-markers");
 
 var markoutMap = null;
 
@@ -34,8 +33,7 @@ function markout(html, markoutMap) {
 
 	cheerioMarkoutMethods($);
 	normalizeMarkers($, markoutMap);
-	missingMarkers($, markoutMap);
-	unknownMarkers($, markoutMap);
+	invalidMarkers($, markoutMap);
     convertMarkers($, markoutMap);
     constructSliders($, markoutMap);
 	constructImages($, markoutMap);
